responding to sonnet comments
