+++ packages/stats/gbstats/models/results.py
@@ -46,11 +46,19 @@ class BaselineResponse:
 
 
 @dataclass
+class PowerResponse:
+    effect_size: float
+    power: float
+    additional_days_needed: float
+
+
+@dataclass
 class BaseVariationResponse(BaselineResponse):
     expected: float
     uplift: Uplift
     ci: Tuple[float, float]
     errorMessage: Optional[str]
+    powerResponse: Optional[PowerResponse]
 
 
 @dataclass
